Added rendering and updated main
Now main supports obstacles in env.
Rendering looks nicer
